requirements.txt
flask_ml==0.2.5
numpy==2.2.2
onnxruntime==1.20.1
Pillow==11.1.0
torch==2.5.1
torchvision==0.20.1

